Fix (RandomGenerator) : fix test codes
